refactor: 상품 뷰 로직 수정
