+++ src/components/products/ProductDetail.tsx
@@ -1,7 +1,12 @@
+import { useDispatch } from "react-redux";
+import { addToCart } from "../../store/cart";
 import Rating from "../common/Rating";
 
 const ProductDetail = ({ product }): JSX.Element => {
   if (product === undefined) return <div></div>;
+
+  const dispatch = useDispatch();
+  console.log(product);
   return (
     <div key={product.id} className="lg:flex lg: gap-36 lg:items-center mt-6 md:mt-14 px-2 lg:px-0">
       <div className="lg:flex lg:items-center mt-6 md:mt-14 px-2 lg:px-0">
@@ -15,11 +20,12 @@ const ProductDetail = ({ product }): JSX.Element => {
           <span className="badge badge-accent ml-2">NEW</span>
         </h2>
         <p className="">{product.description}</p>
-
         <Rating rate={product?.rating.rate} count={product?.rating.count} />
         <p className="mt-2 mb-4 text-3xl">$ {product?.price}</p>
         <div className="card-actions">
-          <button className="btn btn-primary">장바구니에 담기</button>
+          <button className="btn btn-primary" onClick={() => dispatch(addToCart({}))}>
+            장바구니에 담기
+          </button>
           <a className="btn btn-outline ml-1" href="/carts">
             장바구니로 이동
           </a>
